Add Styled-Components to project and refactor Page and Layout components
This commit introduces Styled-Components to the project for a cleaner and more efficient way to write CSS in Javascript. A new types folder and files have been added so that Styled-Components can be typed with Typescript. Default theming with Styled-Components was integrated.

In addition, a refactor of the Page and Layout components was made, removing unnecessary imports and adding a ThemeProvider to the root of the app. Finally, a global styling file and default theme file were created to house all global styles and themes respectively.