"Remove unused _document.tsx and update styles import in layout.tsx"
This commit removes the unused _document.tsx file and updates the way styles are imported in layout.tsx. It also updates the version of several npm packages including '@rocketseat/eslint-config' and 'rimraf'. The change is part of a refactoring process to remove redundant code and improve code clarity and organisation.