Add Main component to the home function, introduce new styles
This commit introduces the Main component, containing the main part of the app, to the home function in `page.tsx` file. New styles have been added as well to `style.css` to further tailor the look of the user interface. The purpose of these changes is to enhance the app's appearance and improve the organization of the components.